refactor: merge skill-* domains + move activity to me [part of #66]
Structural reorg for Epic 1 — resource-oriented modules and caller-scoped
telemetry grouped under the right domain.

Domain layout:
  domains/skillCrud/       → domains/skills/crud/
  domains/skillSearch/     → domains/skills/search/
  domains/skillFormat/     → domains/skills/format/
  domains/skillGeneration/ → domains/skills/generation/

Four verb-oriented siblings now form one resource-oriented skills/ domain
with four submodules. Matches conventions §11.4.

Caller telemetry moved out of admin:
  POST /activity/login  — admin/routes.ts → me/routes.ts
  POST /activity/logout — admin/routes.ts → me/routes.ts

They were never admin operations (any authed user logs their own session
events). The admin domain now only exposes /admin/* gated by admin
permissions. Path unchanged.

Mechanical updates:
- bootstrap.ts: 9 import paths updated to domains/skills/* layout.
- Cross-domain imports (me/, admin/, playground/chatService.ts) point
  to ../skills/crud/*.
- Intra-skills sibling imports: ../skillCrud/* → ../crud/*.
- Every relative import inside skills/* gained one ../ (depth +1).
- @module JSDoc comments updated.

No external /api/* contract change. 136 tests pass. Backend typecheck
has only pre-existing errors (unchanged count).
