+++ ornn-api/src/domains/skills/crud/repositories/skillRepository.interface.ts
