+++ ornn-api/src/domains/skills/generation/types/generation.ts
