+++ ornn-api/src/domains/skills/generation/types/streaming.ts
