+++ ornn-api/src/domains/skills/search/middleware/apiKeyMiddleware.test.ts
@@ -2,8 +2,8 @@ import { describe, test, expect, beforeEach, mock } from "bun:test";
 import { Hono } from "hono";
 import pino from "pino";
 import { createApiKeyMiddleware, getApiKeyUser } from "./apiKeyMiddleware";
-import type { IAuthClient } from "../../../clients/authClient";
-import { createErrorHandler } from "../../../shared/types/index";
+import type { IAuthClient } from "../../../../clients/authClient";
+import { createErrorHandler } from "../../../../shared/types/index";
 
 const silentLogger = pino({ level: "silent" });
 const errorHandler = createErrorHandler(silentLogger);
